Added --rgbxy custom palette support
Added a debug section to add args as though it was executed from a command prompt when debugging with an IDE. Added an automatic file downloader for heatmap.py and flatten.py, and removed them from the git location to avoid duplication. Added support for custom palettes (--palette and --rgbxy). Fixed automation interval to use milliseconds not seconds. Added check to avoid processing csv file if it has not changed since last update.