Changed RGBXY value to be single arg. Added --ratio
Changed the RGBXY formatting to be a single sys.arg value separated by colons (e.g. 000:255:255:015:140). This breaks backwards compatibility. Also added a --ratio argument that allows the size of the spectrum vs. waterfall to be adjusted. Uses two values separated by a colon (e.g. 3:1) First number is total rows, second number is the number of rows allocated for the spectrum. Remaining rows are for the waterfall.